(refact) Split up OverviewTableView code into separate files to improve legibility
