+++ web/src/ResourceStatusSummary.tsx
@@ -7,7 +7,7 @@ import { ReactComponent as PendingSvg } from "./assets/svg/pending.svg"
 import { ReactComponent as WarningSvg } from "./assets/svg/warning.svg"
 import { FilterLevel } from "./logfilters"
 import { useLogStore } from "./LogStore"
-import { RowValues } from "./OverviewTable"
+import { RowValues } from "./OverviewTableColumns"
 import { usePathBuilder } from "./PathBuilder"
 import SidebarItem from "./SidebarItem"
 import { buildStatus, combinedStatus, runtimeStatus } from "./status"
